requirements.txt
matplotlib==3.9.2
numpy==2.1.1
pykan==0.2.6
pysr==0.19.4
sympy==1.11.1
torch
